requirements.txt
# Requirements for macOS

# Core AI
google-generativeai>=0.3.0

# GUI
PyQt5>=5.15.0

# Computer Control
pyautogui>=0.9.53

# Environment
python-dotenv>=0.19.0

# Image Processing
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0

# HTTP Requests (if needed)
requests>=2.26.0

# Protobuf - pin to avoid conflicts with mediapipe
protobuf==3.20.3

# Global keyboard shortcuts
pynput>=1.7.6

# Type hints (Python 3.8+)
typing-extensions>=4.0.0

# Task scheduling for proactive agent
schedule>=1.1.0

# macOS app detection and control
pyobjc-framework-Cocoa>=8.0
pyobjc-core>=8.0
